final changes completed
