+++ src/components/organisms/section-2/Section2.js
@@ -1,82 +1,85 @@
-import { useI18n } from "~/i18n";
+import PropTypes from 'prop-types';
+import { useEffect } from 'react';
+import ImageAtomLocal from '~/components/atoms/imageAtom/ImageAtomLocal';
 import Title from '~/components/atoms/Title/Title';
-import ImageSection2 from '~/assets/static/images/section2/mobilecircle.png';
-import Image from "next/image";
+import Button2 from '~/components/atoms/button/Button1.js';
 
-export default function Section2({
-    text,
-    subtext,
-    title1,
-    sub1,
-    title2,
-    sub2,
-    title3,
-    sub3,
-    disclaimer,
-    buttonText
+export default function Section1({
+  sizeImg,
+  borderImg,
+  src,
+  altImg,
+  text,
+  text2,
+  mainText,
+  subText,
+  buttonText,
+  disclaimer1,
+  disclaimer2,
+  mainText2,
+  subText2
 }) {
-  const { } = useI18n();
-  return (
-    // Black background
-    <div className="flex flex-col items-center justify-center w-full min-h-[1050px] xl:min-h-[1273px] bg-[#0A0A0A] pb-[2rem]">
-      {/* Padding and margin */}
-      <div className="w-full flex items-center justify-center mt-[1rem] px-[20px] xl:px-[50px]">
-        {/* Blue foreground */}
-        <div className="w-full bg-[#132D62] rounded-[50px] flex flex-col items-center pb-[2rem]">
+  useEffect(() => {
+    const style = document.createElement('style');
+    style.innerHTML = `
+      .gradient-text {
+        background: linear-gradient(180deg, #30DA74 100%, #3578DE 100%);
+        -webkit-background-clip: text;
+        -webkit-text-fill-color: transparent;
+      }
+    `;
+    document.head.appendChild(style);
+    
+    return () => {
+      document.head.removeChild(style);
+    };
+  }, []);
 
-          <div className="flex flex-col items-center mt-[2rem] lg:mt-[4rem] w-full px-4 text-center">
+  const applyGradient = (text) => {
+    const gradientText = text.replace(
+      /CoughCheck/g,
+      '<span class="gradient-text">CoughCheck</span>'
+    );
+    return { __html: gradientText };
+  };
+
+  return (
+    <div className="relative flex justify-center overflow-hidden bg-[#0A0A0A]">
+      <div className="contrast-1 brightness-100">
+        <ImageAtomLocal
+          ImageStyleProps=""
+          imagesize={sizeImg}
+          border={borderImg}
+          src={src}
+          alt={altImg}
+        />
+      </div>
+      <div className="absolute inset-0 flex flex-col justify-start p-4 md:p-8 lg:p-12 text-center sm:text-left text-white">
+        <div className="flex flex-col mt-12 text-center xl:mt-[18rem] sm:mx-0 lg:mx-20 ml-[0rem] lg:ml-[6rem]">
+          <Title Text={text} H="" TitleClassProps="text-white text-[0.99rem] sm:leading-[2.25rem] md:text-[2.25rem] md:leading-[2.5rem] lg:text-[2.5rem] lg:leading-[2.75rem] xl:text-[3rem] xl:leading-[3.5rem]"/>
+          <div className="flex flex-col mt-0 lg:mt-12 sm:text-center">
             <Title
-              Text={text}
+              Text={text2}
               H=""
-              TitleClassProps="text-transparent bg-clip-text bg-gradient-to-b from-[#30DA74] to-[#3578DE] text-[20px] leading-[1rem] sm:leading-[2.25rem] md:text-[2.25rem] md:leading-[2.5rem] lg:text-[2.5rem] lg:leading-[2.75rem] xl:text-[48px] xl:leading-[3.5rem]"
+              TitleClassProps="font-bold text-transparent bg-clip-text bg-gradient-to-b from-[#30DA74] to-[#3578DE] text-[0.99rem] sm:leading-[2.25rem] md:text-[2.25rem] md:leading-[2.5rem] lg:text-[2.5rem] lg:leading-[2.75rem] xl:text-[3rem] xl:leading-[3.5rem]"
             />
-            <p className="mt-[1rem] px-[1px] xl:px-[160px] text-[14px] text-white leading-[1rem] md:text-[1.125rem] md:leading-[1.75rem] lg:text-[1.25rem] lg:leading-[2rem] xl:text-[24px] xl:leading-[2.5rem] xl:font-bold">
-              {subtext}
+          </div>
+        </div>
+        <div className="flex flex-col items-start mt-24 lg:mt-100 xl:mt-[27rem] space-y-2 md:space-y-8 sm:ml-0 lg:ml-20">
+          <div className="w-full max-w-xs md:max-w-md">
+            <p className="text-[0.75rem] leading-[1rem] md:text-[1rem] md:leading-[1.5rem] lg:text-[1.125rem] lg:leading-[1.75rem] xl:text-[1.25rem] xl:leading-[2rem]">
+              {mainText}
             </p>
           </div>
 
-          <div className="flex flex-col sm:flex-col xl:flex-row w-full mt-[0rem] xl:mt-[2rem] px-[2rem] text-center xl:text-start xl:px-[100px] justify-between">
-            <div className="order-2 xl:order-1 flex flex-col items-center xl:items-start sm:w-full xl:w-1/2 px-[0rem] xl:px-[6rem] xl:mt-[4rem]">
-              <div className="mt-[1rem] xl:mt-[2rem] mb-[1rem]">
-                <Title H="h5spaced" Text={title1} TitleClassProps="text-white font-bold mb-[1rem]" />
-                <p className="text-white text-[0.9rem] leading-[0.9rem] md:text-[1.25rem] md:leading-[1.5rem] lg:text-[1.5rem] lg:leading-[1.75rem] xl:text-[18px] xl:leading-[25px]">
-                  {sub1}
-                </p>
-              </div>
-              <div className="mt-[0rem] xl:mt-[1rem] mb-[1rem]">
-                <Title H="h5spaced" Text={title2} TitleClassProps="text-white font-bold mb-[1rem]" />
-                <p className="text-white text-[0.9rem] leading-[0.9rem] md:text-[1.25rem] md:leading-[1.5rem] lg:text-[1.5rem] lg:leading-[1.75rem] xl:text-[18px] xl:leading-[25px]">
-                  {sub2}
-                </p>
-              </div>
-              <div className="mt-[1rem] xl:mt-[1rem] mb-[0rem]">
-                <Title H="h5spaced" Text={title3} TitleClassProps="text-white font-bold mb-[1rem] whitespace-nowrap" />
-                <p className="text-white text-[0.9rem] leading-[0.9rem] md:text-[1.25rem] md:leading-[1.5rem] lg:text-[1.5rem] lg:leading-[1.75rem] xl:text-[18px] xl:leading-[25px]">
-                  {sub3}
-                </p>
-              </div>
-              <div className="flex flex-col items-center xl:items-start mt-[2rem] xl:mt-[4rem] xl:whitespace-nowrap w-full">
-                <div className="w-full">
-                  <p className="text-[0.4rem] px-[3rem] xl:px-[0rem] md:text-[0.875rem] lg:text-[1rem] xl:text-[8px] text-white text-center xl:text-left">
-                    {disclaimer}
-                  </p>
-                </div>
-              </div>
-            </div>
-            <div className="order-1 xl:order-2 flex justify-center xl:w-1/2 mt-[2rem] xl:mt-[0rem]">
-              <div className="w-auto relative">
-                <Image
-                  className="w-auto h-[300px] xl:h-[800px]"
-                  src={ImageSection2}
-                  width="0"
-                  height="0"
-                  alt="Image of a mobile with Virufy's logo on it"
-                />
-              </div>
-            </div>
+          <div className="w-full max-w-md md:max-w-lg">
+            <p
+              className="text-[0.7rem] leading-[0.9rem] md:text-[1.125rem] md:leading-[1.75rem] lg:text-[1.25rem] lg:leading-[2rem] xl:text-[1.5rem] xl:leading-[2.5rem] font-bold"
+              dangerouslySetInnerHTML={applyGradient(subText)}
+            />
           </div>
-
-          <div className="w-full flex justify-center xl:justify-start max-w-md md:max-w-lg items-center px-0 mt-[2rem] xl:mt-[0rem] mb-[2rem]">
+          
+          <div className="w-full max-w-md md:max-w-lg px-0 mt-2">
             <button
                 className={`medium primary text-white
                             w-[315px] h-[45px]
@@ -89,7 +92,29 @@ export default function Section2({
               >
                 {buttonText}
               </button>
+              
           </div>
+        </div>
+        <div className="flex flex-col items-start mt-1 xl:mt-4 sm:ml-0 lg:ml-20">
+          <div className="w-full max-w-screen-lg">
+            <p className="text-[0.4rem] leading-[0.6rem] md:text-[0.875rem] md:leading-[1.25rem] lg:text-[1rem] lg:leading-[1.5rem] xl:text-[1.125rem] xl:leading-[1.75rem] text-gray-400">
+              {disclaimer1}
+            </p>
+          </div>
+          <div className="w-full max-w-screen-lg">
+            <p className="text-[0.4rem] leading-[0.6rem] md:text-[0.875rem] md:leading-[1.25rem] lg:text-[1rem] lg:leading-[1.5rem] xl:text-[1.125rem] xl:leading-[1.75rem] text-gray-400">
+              {disclaimer2}
+            </p>
+          </div>
+        </div>
+        <div className="bottom-[2rem] md:bottom-[8rem] xl:bottom-[12rem] px-[2rem] absolute flex flex-col items-center mt-[1rem] sm:mt-[2rem] lg:mt-[4rem] xl:mt-[15rem] md:space-y-8 sm:space-y-2 ml-[7rem] lg:ml-[35rem]">
+          <p className="text-[0.8rem] leading-[1.5rem] md:text-[1.5rem] md:leading-[2rem] lg:text-[1.75rem] lg:leading-[2.25rem] xl:text-[2rem] xl:leading-[2.5rem]">
+           {mainText2}
+          </p>
+          <p
+            className="text-[0.5rem] text-center leading-[0.8rem] md:text-[1.125rem] md:leading-[1.75rem] lg:text-[1.25rem] lg:leading-[2rem] xl:text-[1.5rem] xl:leading-[3rem] font-bold"
+            dangerouslySetInnerHTML={applyGradient(subText2)}
+          />
         </div>
       </div>
     </div>
